+++ app/Http/Resources/SummonerRanked.php
@@ -21,7 +21,7 @@ class SummonerRanked extends JsonResource
             'avg gold at 10m' => self::readable($this->gold_at_10),
             'avg game duration' => gmdate("H:i:s", $this->game_duration),
             'avg KDA' => self::readable($this->kda),
-            'winrate %' => self::readable($this->winrate*100)
+            'winrate %' => (float)self::readable($this->winrate*100)
         ];
     }
